backend/requirements.txt
Flask
azure-mgmt-containerinstance
azure-identity
dotenv